glasses/requirements.txt
requests
python-dotenv
luma.oled
pillow
smbus2
picamzero
pyserial
pynmea2
numpy
sounddevice
faster-whisper
pexpect
